requirements.txt
flask>=3.0
flask-login>=0.6
flask-sqlalchemy>=3.1
python-dotenv>=1.0
